Reorganize schema classes into separate modules
