+++ src/marshmallow_sqlalchemy/schema/__init__.py
@@ -1,0 +1,11 @@
+from .model_schema import ModelSchema, ModelSchemaOpts, ModelSchemaMeta
+from .table_schema import TableSchema, TableSchemaOpts, TableSchemaMeta
+
+__all__ = [
+    "ModelSchema",
+    "ModelSchemaOpts",
+    "ModelSchemaMeta",
+    "TableSchema",
+    "TableSchemaOpts",
+    "TableSchemaMeta",
+]
